app_examples/intelligent_doctor/surgical_agent/pyproject.toml
[project]
name = "cybertwin-surgical-agent"
version = "0.1.0"
description = "外科医生智能体"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.15",
    "aiohttp==3.9.3",
    "aiomysql>=0.2.0",
    "bcrypt==4.1.2",
    "cryptography==41.0.7",
    "dbutils==3.0.3",
    "fastapi==0.110.0",
    "fastmcp>=0.1.0",
    "httpx>=0.24.0",
    "itsdangerous>=2.2.0",
    "mcp>=0.1.0",
    "nest-asyncio==1.6.0",
    "passlib==1.7.4",
    "pillow>=10.2.0",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.6.3",
    "pyjwt==2.8.0",
    "pymysql>=1.1.0",
    "pypdf2==3.0.1",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.9",
    "requests==2.31.0",
    "sqlalchemy==2.0.28",
    "sse-starlette==1.6.5",
    "structlog>=25.5.0",
    "uvicorn==0.27.1",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
    "factory-boy>=3.3.0",
]
